pyproject.toml
[project]
name = "FIIT"
version = "0.1.0"
description = "Интерактивная панель для мониторинга и торговли валютами"
authors = [
    {name = "eldar",email = "eldar.salm@gmail.com"},
    {name = "mikhail",email = "misha.sedov07@gmail.com"},
    {name = "kirill",email = "belianskiykirill@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt[token-blacklist] (>=5.5.0,<6.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "psycopg-binary (>=3.2.6,<4.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
